express 08

+++ userHandlers.js
@@ -26,6 +26,25 @@ const getUsers = (req, res) => {
     });
 };
 
+const getUserByEmailWithPasswordAndPassToNext = (req, res, next) => {
+  const { email } = req.body;
+  database
+    .query("select * from users where email = ?", [email])
+    .then(([users]) => {
+      if (users[0] != null) {
+        req.user = users[0];
+        next();
+      } else {
+        res.status(401);
+      }
+    })
+    .catch((err) => {
+      console.error(err);
+      res.status(500).send("Error retrieving data from database");
+    });
+};
+
+
 const getUserById = (req, res) => {
   const id = parseInt(req.params.id);
 
@@ -70,6 +89,10 @@ const getUserById = (req, res) => {
 
 const updateUser = (req, res) => {
   const id = parseInt(req.params.id);
+  if (req.payload.sub !== id) {
+    res.sendStatus(403);
+    return;
+  }
   const { firstname, lastname, email, city, language } = req.body;
   database
     .query(
@@ -91,6 +114,10 @@ const updateUser = (req, res) => {
 
 const deleteUser = (req, res) => {
   const id = parseInt(req.params.id);
+  if (req.payload.sub !== id) {
+    res.sendStatus(403);
+    return;
+  }
   database
     .query("delete from users where id = ?", [id])
     .then(([result]) => {
@@ -106,11 +133,11 @@ const deleteUser = (req, res) => {
     });
 };
 
-
 module.exports = {
 getUsers,
 getUserById,
 postUsers,
 updateUser,
-deleteUser,  
+deleteUser,
+getUserByEmailWithPasswordAndPassToNext,  
 };
